Separate DataSourceSemantics / MetricSemantics.
