+++ metricflow/model/semantic_model.py
@@ -1,6 +1,7 @@
 from metricflow.model.objects.user_configured_model import UserConfiguredModel
 from metricflow.model.semantics.data_source_container import PydanticDataSourceContainer
-from metricflow.model.semantics.semantic_containers import DataSourceSemantics, MetricSemantics
+from metricflow.model.semantics.data_source_semantics import DataSourceSemantics
+from metricflow.model.semantics.metric_semantics import MetricSemantics
 from metricflow.protocols.semantics import DataSourceSemanticsAccessor, MetricSemanticsAccessor
