+++ metricflow/plan_conversion/node_processor.py
@@ -13,7 +13,7 @@ from metricflow.dataflow.dataflow_plan import (
     JoinDescription,
 )
 from metricflow.model.objects.elements.identifier import IdentifierType
-from metricflow.model.semantics.semantic_containers import MAX_JOIN_HOPS
+from metricflow.model.semantics.metric_semantics import MAX_JOIN_HOPS
 from metricflow.object_utils import pformat_big_objects
 from metricflow.plan_conversion.sql_dataset import SqlDataSet
 from metricflow.protocols.semantics import DataSourceSemanticsAccessor
